Replaced key structs in separate file
