+++ kuzmina.daria/T2/DelimiterI.cpp
@@ -1,7 +1,6 @@
 #include "DelimiterI.hpp"
 #include <istream>
 #include <cctype>
-#include <string>
 
 std::istream& kuzmina::operator>>(std::istream& in, DelimiterI&& exp)
 {
@@ -33,64 +32,6 @@ std::istream& kuzmina::operator>>(std::istream& in, DelimiterIStr&& exp)
   for (size_t i = 0; exp.expected[i] != '\0'; ++i)
   {
     in >> DelimiterI{ exp.expected[i] };
-  }
-
-  return in;
-}
-
-std::istream& kuzmina::operator>>(std::istream& in, SLLKey&& key)
-{
-  std::istream::sentry guard(in);
-  if (!guard)
-  {
-    return in;
-  }
-
-  long long temp = 0;
-  in >> temp >> DelimiterIStr{ "ll" };
-
-  if (in)
-  {
-    key.value = temp;
-  }
-
-  return in;
-}
-
-std::istream& kuzmina::operator>>(std::istream& in, RATKey&& key)
-{
-  std::istream::sentry guard(in);
-  if (!guard)
-  {
-    return in;
-  }
-
-  long long tempN = 0;
-  unsigned long long tempD = 0;
-  in >> DelimiterIStr{ "(:N" } >> tempN >> DelimiterIStr{ ":D" } >> tempD >> DelimiterIStr{ ":)" };
-
-  if (in)
-  {
-    key.value = { tempN, tempD };
-  }
-
-  return in;
-}
-
-std::istream& kuzmina::operator>>(std::istream& in, STRKey&& key)
-{
-  std::istream::sentry guard(in);
-  if (!guard)
-  {
-    return in;
-  }
-
-  std::string tempStr = "";
-  std::getline(in >> DelimiterI{ '"' }, tempStr, '"');
-
-  if (in)
-  {
-    key.value = tempStr;
   }
 
   return in;
